chore(allsrv): adding test for update foo API
The key takeaway here is the response is a nothing burger. We can only ever
test the status code atm. These leaves a lot to be desired. As a service
evolves to cover a broader domain, these types tend to grow. We don't see
anything in an empty response however.

At the moment we're kind of stuck though. We don't have a means to add a
new API that is distinguished from the existing. We'll come back to this.