chore(allsrv): add database observer for DB metrics
Often times, metrics and tracing are left out, leaving the service
owners blind. When we add metrics, with defined patterns, we're able to
build out robust metrics and tracing dashboards.

Additionally, with the observer we get a warning from the compiler
that we forgot to add observability concerns for any new behavior/method
added to the database(s).

Now, how about adding opentracing spans to the observer?