chore(allsrv): add tracing to the database observer
The observer is updated with additional observability concerns. This
does violate single responsibility principles, but in this case it
encapsulates well. In my experience, it's rare that you have
metrics and tracing concerns that are required, and you only want one
or the other. Having a single observer for these makes it fairly simple.

Now that these are in place, add observability for metrics/tracing to
the http server.