chore(allsrv): add http server observability
The HTTP server now has quantification for different metrics important
to an HTTP server. The basis of our observability is now in place. We
can now create dashboards/insights to understand the deployed service.

One thing to note here is we have not touched on logging just yet. Good
logging is inherently coupled to good error handling. We'll wait until
we have a better handle of our error handling before proceeding.