chore(allsrv): add tests for the in-mem db
This helps us close the gap in our testing. This time we're putting the
in-mem db under test. This is partially under test via the server tests,
but we have limited visibility into the stack. Once we have the basics
in place, we can start to ask more interesting quetsions of our system.

Try to create a test that will trigger the race condition in the in-mem
operations for each destructive operation? Hint: use the `-race` flag:

```shell
go test -race
```