fix(allsrv): serialize access for in-mem value access to rm race condition
This is pretty straight forward. We just add the `-race` flag to our `go test`
invocation after we add some tests that access the db concurrently. There
are many more test cases to add for this specific instance, but the point
is made with the existing one.

We're starting to get somewhat comfortable with our existing test suite. If
this were a real world service, you'd have tests at a higher level and more
integration and end to end testing. These tests are slow but are the most
valuable as they are validating more integration points! Unit tests can easily create a
false sense of safety. Especially when unit testing that is MOCKING ALL
THE THINGS!

Note, we can use other mechanisms to address the race condition. However,
it gets complex... FAST. Katherine Cox-Buday's "Concurrency in Go" is a
wonderful deep dive on the subject :-):

Refs: [Concurrency in Go - Katherine Cox-Buday](https://katherine.cox-buday.com/concurrency-in-go/)