chore(allsrv): add structured errors
Structured errors are so incredibly helpful for service owners. With
structured errors, you're able to enrich the context of your failures
without having to add DEBUG log after DEBUG log, to provide that context.

This marks the end of the "training wheels" part of the workshop. Now that
we have a somewhat solid foundation to build on, we can start to make
informed decisions that help us survive the only constant in software
engineering... changing business requirements >_<!

We still have a number of concerns that we can't address within the
the existing API deisgn... but now we have the tools to inform
our server evolution. GEDIUP!