chore(allsrv): add db test suite
By adding a DB test suite, we've effectively lowered the bar to entry for
any additional DBs we wish to add. With this test suite, we can re-run it
against a SQL DB, nosql db, etc. and expect it to satisfy the behavior the
tests cover. Interestingly, this is not limited to dbs. There are a number
of great examples of test suites in open source software. Here are a couple
examples:

  * [influxdb](https://github.com/influxdata/influxdb/tree/v2.0.0/testing):
              things to note here are the abstraction around KV and the
              service behavior. However, reusing the pkg name, testing,
              is not something i'd advise doing. It will create confusion.
  * [vice](https://github.com/matryer/vice/blob/master/vicetest/test.go):
          this lib does a great job abstracting over the expected _behavior_
          similar to influxdb. When you look through the test for each queue
          you should see the `vicetest` pkg's exported tests being called.
          The pkg name, vicetest, is very explicit, would highly recommend
          using a similar naming strategy.

The key thing here is the language of the database interface. Just as
the two examples above are abstracting over the behavior, we do the same
here. Since the closest thing we have to a domain type is the `Foo` type,
we utilize that as our domain language. Any db will need to be able take
a domain Foo and persist it in using whatever implementation they desire. We aren't
bleeding any details beyond the point of implementation. To illustrate this,
I removed the GORM struct tags, as the domain type should not be limited
by the db design.

Now that we have a little test suite stood up, go on and add a new db implementation
and make sure it passes the test suite. I will be adding a sql store, with
sqlite, as a example to build from, however, feel free to explore this problem
space however you'd like.

Refs: [InfluxDB Testing](https://github.com/influxdata/influxdb/tree/v2.0.0/testing)
Refs: [Vice Test Suite](https://github.com/matryer/vice/blob/master/vicetest/test.go)