chore(allsrv): add sqlite db implementation and test suite
The test suite underwent a few changes here. There are race conditions
in the original test suite, when it comes to executing the concurrency
focused tests. Made a small update here to address that. TL/DR the race is with
the actual `*testing.T` type, so we make use of the closure to capture the
error and log. This enforces the `testing.T` access is done **after** any test
behavior.

We are also able to update the error handling a bit here. I don't care much
for what the error message says, but I care deeply about the behavior of
the errors I receive. I want to validate the correct behavior is obtained.
This is very useful when integrating amongst a larger, more complex system.
For this trivial `Foo` server, we don't have much complexity.

The sqlite db implementation here is fairly trivial once again. We're able
to reuse the test suite in full. All that was required is a new funciton to initize the unit under test (UUT),
and the rest is including 4 more lines of code to call the test around it.
Not to shabby.

Since we've effectively decoupled our domain `Foo` from the db entity `foo`,
we've provided maximum flexibility to our database implementation without
having to pollute the domain space. This is intensely useful as a system
grows.

Think through, what would it look like to add a `PostgreSQL` db implementation?
Not as much now that you have a test suite to verify the desired behavior.

The last thing that is missing here is what we do to decouple our server
from HTTP. There is a glaring hole in our design, and that's the lack of
service layer. The layer where all our business logic resides. Take a moment
to think through what this might look like.

Question: How would you break up the server so that it's no longer coupled to HTTP/REST?
Question: What can we do to allow ourselves to support any myriad of `RPC` technologies without duplicating allthe business logic?