chore(allsrv): add the service layer to consolidate all domain logic
This might seem like a "moving the cheese" change. However, upon closer look
we see that the `server_v2` implementation is purely a translation between
the HTTP RESTful API and the domain. All traffic speaks to the service,
which holds all the logic for the Foo domain.

We've effectively decoupled the domain from the transport layer (HTTP).
Any additional transport we want to support (gRPC/Thrift/etc) is merely
creating the transport implementation. We won't duplicate our logic in each transport layer.
Often, when we have consolidated all the business logic, it's very simple to just generate the RPC layer and inject
the SVC to transact with the different API integrations.