chore(allsrv): provide test suite for SVC behavior and fill gaps in behavior
This test suite provides the rest of the codebase super powers. When we have
new requirements to add, we can extend the service's test suite and we can use
that implementation across any number of implementation details. However,
the super power shows up when we start to integrate multiple service
implementations.

Here's a thought exercise:

Perhaps there is a `Bar` service that integrates the `Foo` service as part of a modular
monolith design. Now you have scale/requirements hitting you that force you
to scale `Foo` independent of `Bar` or vice versa. Perhaps we pull out the foo
svc into its own deployment. Now our `Bar` service needs to access the `Foo` service
via some RPC channel (HTTP|REST/gRPC/etc.).

We then create a remote SVC implementation, perhaps an HTTP client, that
implements the SVC behavior (interface). How do we verify this adheres to
the service behavior? Simple enough, just create another test with the
service's test suite, initilaize the necessary components, and excute them...
pretty simple and guaranteed to be in line with the previous implementation.

Question: Now that we have a remote SVC implementation via an HTTP client, what else
might we want to provide?
Question: How about a CLI that can integrate with the `Foo`?
Question: How would we test a CLI to validate it satisfies the `SVC` interface?