chore(allsrv): add http client and fixup missing pieces
Here we've added the HTTP client. Again, we're pulling from the standard library
because it's a trivial example. Even with this, we're able to put together a client
that speaks the languae of our domain, and fulfills the behavior of our SVC.
We've provided a fair degree of confidence by utilizing the same `SVC` test
suite we had with the `Service` implementation itself. To top it all off,
we're able to refactor our tests a bit to reuse the constructor for the
`SVC` dependency, leaving us with a standardized setup.

With standardized tests you benefit of reusing the tests. Additionally,
any new contributor only needs to understand a single test setup,
and then writes a testcase. Its extremely straightforward after the
initial onboarding.

Last commit message, I spoke of adding a CLI companion to this. With the HTTP
client we've just created, go on and create a CLI and put it under test
with the same test suite :yaaaaaaaaaas:!