chore(allsrv): add allsrvc CLI companion
Again here we start with a little refactoring. This time we're creating
a new `allsrvtesting` pkg to hold the reusable code bits. Now we can
implement our CLI and then create the svc implemenation utilizing the cli
in our tests. With a handful of lines of code we're able to create a
high degree of certainty in the implementation of our CLI. The
implementation and testing both, are not limited by the SVC behavior.
You can extend the CLI with additional behavior beyond that of the SVC.
Additional tests can be added to accomodate any additional behavior.

Here we are at the end of the session where we've matured an intensely
immature service implementation. We've covered a lot of ground. We can
sum it up quickly with:

  1. Tests provide certainty we've not broken the existing implementation
  2. Versioning an API helps us transition into the replacement
      * note: we determined in that the v1 was not serving our users
              best interest, so we moved onto a structured JSON-API spec
  3. Observability is SUUUPER important. Just like with testing, we want
     to keep a close eye on our metrics. We want to make sure our changes
     are improving the bottom line. Without this information, we're flying
     blind.
  4. Isolating the SVC/usecases from the transport & db layers gives us
     freedom to reuse that business logic across any number of transport
     & db layers, improve our observability stack along the way, and allow
     us to create a reusable test suite that is usable across any
     implementation of the SVC. With that test suite, creating and
     verifying a new SVC implementation.